--- a/src/batch.jl
+++ b/src/batch.jl
@@ -1,27 +1,38 @@
-struct BatchClosure{F,A,C} # C is a Val{Bool} triggering local storage
-  f::F
+# S is a Val{Bool} indicating whether we will need to load the thread index
+# C is a Tuple{...} containing the types of the reduction variables
+struct BatchClosure{F,A,S,C}
+    f::F
 end
-function (b::BatchClosure{F,A,C})(p::Ptr{UInt}) where {F,A,C}
+function (b::BatchClosure{F,A,S,C})(p::Ptr{UInt}) where {F,A,S,C}
   (offset, args) = ThreadingUtilities.load(p, A, 2 * sizeof(UInt))
   (offset, start) = ThreadingUtilities.load(p, UInt, offset)
   (offset, stop) = ThreadingUtilities.load(p, UInt, offset)
-  if C
+  if C === Tuple{} && !S
+    b.f(args, (start + one(UInt)) % Int, stop % Int)
+  elseif C === Tuple{} && S
     ((offset, i) = ThreadingUtilities.load(p, UInt, offset))
     b.f(args, (start + one(UInt)) % Int, stop % Int, i % Int)
+  elseif C !== Tuple{} && !S
+    ((offset, reducinits) = ThreadingUtilities.load(p, C, offset))
+    reducres = b.f(args, (start + one(UInt)) % Int, stop % Int, reducinits)
+    ThreadingUtilities.store!(p, reducres, offset)
   else
-    b.f(args, (start + one(UInt)) % Int, stop % Int)
+    ((offset, i) = ThreadingUtilities.load(p, UInt, offset))
+    ((offset, reducinits) = ThreadingUtilities.load(p, C, offset))
+    reducres = b.f(args, (start + one(UInt)) % Int, stop % Int, i % Int, reducinits)
+    ThreadingUtilities.store!(p, reducres, offset)
   end
   ThreadingUtilities._atomic_store!(p, ThreadingUtilities.SPIN)
   nothing
 end
 
-@generated function batch_closure(f::F, args::A, ::Val{C}) where {F,A,C}
+@generated function batch_closure(f::F, args::A, ::Val{S}, reducinits::C) where {F,A,S,C}
   q = if Base.issingletontype(F)
-    bc = BatchClosure{F,A,C}(F.instance)
+    bc = BatchClosure{F,A,S,C}(F.instance)
     :(@cfunction($bc, Cvoid, (Ptr{UInt},)))
   else
     quote
-      bc = BatchClosure{F,A,C}(f)
+      bc = BatchClosure{F,A,S,C}(f)
       @cfunction($(Expr(:$, :bc)), Cvoid, (Ptr{UInt},))
     end
   end
@@ -32,6 +43,20 @@ end
 #   @cfunction($bc, Cvoid, (Ptr{UInt},))
 # end
 
+@inline function load_threadlocals(tid, argtup::A, ::Val{S}, reductup::C) where {A,S,C}
+  p = ThreadingUtilities.taskpointer(tid)
+  (offset, _) = ThreadingUtilities.load(p, UInt, sizeof(UInt))
+  (offset, _) = ThreadingUtilities.load(p, A, offset)
+  (offset, _) = ThreadingUtilities.load(p, UInt, offset)
+  (offset, _) = ThreadingUtilities.load(p, UInt, offset)
+  if S
+    (offset, _) = ThreadingUtilities.load(p, UInt, offset)
+  end
+  (offset, _) = ThreadingUtilities.load(p, C, offset)
+  (offset, reducvals) = ThreadingUtilities.load(p, C, offset)
+  return reducvals
+end
+
 @inline function setup_batch!(
   p::Ptr{UInt},
   fptr::Ptr{Cvoid},
@@ -43,6 +68,21 @@ end
   offset = ThreadingUtilities.store!(p, argtup, offset)
   offset = ThreadingUtilities.store!(p, start, offset)
   offset = ThreadingUtilities.store!(p, stop, offset)
+  nothing
+end
+@inline function setup_batch!(
+  p::Ptr{UInt},
+  fptr::Ptr{Cvoid},
+  argtup,
+  start::UInt,
+  stop::UInt,
+  i_or_reductup,
+)
+  offset = ThreadingUtilities.store!(p, fptr, sizeof(UInt))
+  offset = ThreadingUtilities.store!(p, argtup, offset)
+  offset = ThreadingUtilities.store!(p, start, offset)
+  offset = ThreadingUtilities.store!(p, stop, offset)
+  offset = ThreadingUtilities.store!(p, i_or_reductup, offset)
   nothing
 end
 @inline function setup_batch!(
@@ -52,12 +92,14 @@ end
   start::UInt,
   stop::UInt,
   i::UInt,
+  reductup,
 )
   offset = ThreadingUtilities.store!(p, fptr, sizeof(UInt))
   offset = ThreadingUtilities.store!(p, argtup, offset)
   offset = ThreadingUtilities.store!(p, start, offset)
   offset = ThreadingUtilities.store!(p, stop, offset)
   offset = ThreadingUtilities.store!(p, i, offset)
+  offset = ThreadingUtilities.store!(p, reductup, offset)
   nothing
 end
 @inline function launch_batched_thread!(cfunc, tid, argtup, start, stop)
@@ -66,7 +108,20 @@ end
     setup_batch!(p, fptr, argtup, start, stop)
   end
 end
-@inline function launch_batched_thread!(cfunc, tid, argtup, start, stop, i)
+@inline function launch_batched_thread!(cfunc, tid, argtup, start, stop, i_or_reductup)
+  fptr = Base.unsafe_convert(Ptr{Cvoid}, cfunc)
+  ThreadingUtilities.launch(
+    tid,
+    fptr,
+    argtup,
+    start,
+    stop,
+    i_or_reductup,
+  ) do p, fptr, argtup, start, stop, i_or_reductup
+    setup_batch!(p, fptr, argtup, start, stop, i_or_reductup)
+  end
+end
+@inline function launch_batched_thread!(cfunc, tid, argtup, start, stop, i, reductup)
   fptr = Base.unsafe_convert(Ptr{Cvoid}, cfunc)
   ThreadingUtilities.launch(
     tid,
@@ -75,8 +130,9 @@ end
     start,
     stop,
     i,
-  ) do p, fptr, argtup, start, stop, i
-    setup_batch!(p, fptr, argtup, start, stop, i)
+    reductup
+  ) do p, fptr, argtup, start, stop, i, reductup
+    setup_batch!(p, fptr, argtup, start, stop, i, reductup)
   end
 end
 _extract_params(::Type{T}) where {T<:Tuple} = T.parameters
@@ -111,7 +167,9 @@ end
 
 @generated function _batch_no_reserve(
   f!::F,
-  threadlocal::Val{thread_local},
+  needtid::Val{S},
+  reducops::Tuple{Vararg{Any,C}},
+  reducinits::Tuple{Vararg{Any,C}},
   threadmask_tuple::NTuple{N},
   nthread_tuple,
   torelease_tuple,
@@ -119,7 +177,7 @@ end
   Nd,
   ulen,
   args::Vararg{Any,K},
-) where {F,K,N,thread_local}
+) where {F,K,N,S,C}
   q = quote
     $(Expr(:meta, :inline))
     # threads = UnsignedIteratorEarlyStop(threadmask, nthread)
@@ -127,16 +185,54 @@ end
     # nthread_total = sum(nthread_tuple)
     Ndp = Nd + one(Nd)
   end
-  launch_quote = if thread_local
-    :(launch_batched_thread!(cfunc, tid, argtup, start, stop, i % UInt))
+  C !== 0 && push!(
+    q.args,
+    quote
+      @nexprs $C j -> RVAR_j = reducinits[j]
+    end
+  )
+  launch_quote = if S
+    if C === 0
+      :(launch_batched_thread!(cfunc, tid, argtup, start, stop, i % UInt))
+    else
+      :(launch_batched_thread!(cfunc, tid, argtup, start, stop, i % UInt, reducinits))
+    end
   else
-    :(launch_batched_thread!(cfunc, tid, argtup, start, stop))
+    if C === 0
+      :(launch_batched_thread!(cfunc, tid, argtup, start, stop))
+    else
+      :(launch_batched_thread!(cfunc, tid, argtup, start, stop, reducinits))
+    end
   end
-  rem_quote = if thread_local
-    :(f!(arguments, (start + one(UInt)) % Int, ulen % Int, (sum(nthread_tuple) + 1) % Int))
+  f_quote = Expr(:call, :f!, :arguments, :((start + one(UInt)) % Int), :(ulen % Int))
+  S && push!(f_quote.args, :((sum(nthread_tuple) + 1) % Int))
+  C !== 0 && push!(f_quote.args, :reducinits)
+  rem_quote = Expr(:block, :(thread_results = $f_quote))
+  if C !== 0
+    push!(
+      rem_quote.args,
+      :(@nexprs $C j -> RVAR_j = reducops[j](RVAR_j, thread_results[j]))
+    )
+  end
+  update_retv = if C === 0
+      Expr(:block)
   else
-    :(f!(arguments, (start + one(UInt)) % Int, ulen % Int))
+    quote
+      thread_results = load_threadlocals(tid, argtup, needtid, reducinits)
+      @nexprs $C j -> RVAR_j = reducops[j](RVAR_j, thread_results[j])
+    end
   end
+  ret_quote = Expr(:return)
+  if C === 0
+    push!(ret_quote.args, nothing)
+  else
+    redtup = Expr(:tuple)
+    for j in 1:C
+      push!(redtup.args, Symbol("RVAR_", j))
+    end
+    push!(ret_quote.args, redtup)
+  end
+
   block = quote
     start = zero(UInt)
     tid = 0x00000000
@@ -161,17 +257,18 @@ end
     for (threadmask, nthread) ∈ zip(threadmask_tuple, nthread_tuple)
       tm = mask(UnsignedIteratorEarlyStop(threadmask, nthread))
       while tm ≠ zero(tm)
-        # assume(tm ≠ zero(tm)) 
+        # assume(tm ≠ zero(tm))
         tz = trailing_zeros(tm) % UInt32
         tz += 0x00000001
         tm >>>= tz
         tid += tz
         # @show tid, ThreadingUtilities._atomic_state(tid)
         ThreadingUtilities.wait(tid)
+        $update_retv
       end
     end
     free_threads!(torelease_tuple)
-    nothing
+    $ret_quote
   end
   gcpr = Expr(:gc_preserve, block, :cfunc)
   argt = Expr(:tuple)
@@ -182,10 +279,9 @@ end
     q.args,
     :(arguments = $argt),
     :(argtup = Reference(arguments)),
-    :(cfunc = batch_closure(f!, argtup, Val{$thread_local}())),
+    :(cfunc = batch_closure(f!, argtup, Val{$S}(), reducinits)),
     gcpr,
   )
-  push!(q.args, nothing)
   q
 end
 
@@ -195,19 +291,24 @@ end
   args::Vararg{Any,K},
 ) where {F,K}
 
-  batch(f!, Val{false}(), (len, nbatches), args...)
+  batch(f!, Val{false}(), (), (), (len, nbatches), args...)
 end
 
 @inline function batch(
   f!::F,
-  threadlocal::Val{thread_local},
+  needtid::Val{S},
+  reducops::Tuple{Vararg{Any,C}},
+  reducinits::Tuple{Vararg{Any,C}},
   (len, nbatches)::Tuple{Vararg{Union{StaticInt,Integer},2}},
   args::Vararg{Any,K},
-) where {F,K,thread_local}
-  len > 0 || return
+) where {F,K,C,S}
+  len > 0 || return reducinits
+  for var in reducinits
+    @assert isbits(var)
+  end
   if (nbatches > len)
     if (typeof(nbatches) !== typeof(len))
-      return batch(f!, threadlocal, (len, len), args...)
+      return batch(f!, reducops, reducinits, needtid, (len, len), args...)
     end
     nbatches = len
   end
@@ -218,19 +319,32 @@ end
   nthread = sum(nthreads)
   if nthread % Int32 ≤ zero(Int32)
     @label SERIAL
-    if thread_local
-      f!(args, one(Int), ulen % Int, 1)
+    if S
+      if C === 0
+        reducres = f!(args, one(Int), ulen % Int, 1)
+        return reducres
+      else
+        reducres = f!(args, one(Int), ulen % Int, 1, reducinits)
+        return reducres
+      end
     else
-      f!(args, one(Int), ulen % Int)
+      if C === 0
+        reducres = f!(args, one(Int), ulen % Int)
+        return reducres
+      else
+        reducres = f!(args, one(Int), ulen % Int, reducinits)
+        return reducres
+      end
     end
-    return
   end
   nbatch = nthread + one(nthread)
   Nd = Base.udiv_int(ulen, nbatch % UInt) # reasonable for `ulen` to be ≥ 2^32
   Nr = (ulen - Nd * nbatch) % Int
   _batch_no_reserve(
     f!,
-    threadlocal,
+    needtid,
+    reducops,
+    reducinits,
     map(mask, threads),
     nthreads,
     torelease,
@@ -244,7 +358,9 @@ function batch(
   f!::F,
   (len, nbatches, reserve_per_worker)::Tuple{Vararg{Union{StaticInt,Integer},3}},
   args::Vararg{Any,K};
-  threadlocal::Val{thread_local} = Val(false),
-) where {F,K,thread_local}
-  batch(f!, threadlocal, (len, nbatches), args...)
+  needtid::Val{S} = Val(false),
+  reducops::Tuple{Vararg{Any,C}} = (),
+  reducinits::Tuple{Vararg{Any,C}} = (),
+) where {F,K,C,S}
+  batch(f!, needtid, reducops, reducinits, (len, nbatches), args...)
 end