@Krastanov 2024-02-24
My apologies, the readme has been simplified significantly since the last time I looked. Disregard my comment.